+++ src/components/MainPages/ChartPage.tsx
@@ -8,8 +8,8 @@ import { Suspense } from "react";
 
 
 
-const LineChart = React.lazy(() => import("../ChartLine/LineChartComponents"));
-const PieChart = React.lazy(() => import("../ChartPie/PieChartComponents"));
+const LineChart = React.lazy(() => import("../ChartLine/LineChartComponent"));
+const PieChart = React.lazy(() => import("../ChartPie/PieChartComponent"));
 
 
 const monthsShort = ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Okt", "Nov", "Des"];
