+++ src/components/MainPages/wrapper.tsx
@@ -1,6 +1,3 @@
-import React, { Suspense } from "react";
-import LoadingAnimation from "../Loading/LoadingAnimation";
-
 import MainPage from "./index";
 
 export default function WrapperPage() {
