+++ src/pages/loginPage/index.astro
@@ -1,5 +1,5 @@
 ---
-import LoginFormComponent from "../../components/LoginForm/LoginFormComponents";
+import LoginFormComponent from "../../components/LoginForm/LoginFormComponent";
 import Layout from "../../layouts/Layout.astro";
 ---
 <Layout>
